Move shorten_date to datetime helpers
