+++ freqtrade/util/__init__.py
@@ -1,4 +1,5 @@
-from freqtrade.util.datetime_helpers import dt_floor_day, dt_from_ts, dt_now, dt_ts, dt_utc
+from freqtrade.util.datetime_helpers import (dt_floor_day, dt_from_ts, dt_now, dt_ts, dt_utc,
+                                             shorten_date)
 from freqtrade.util.ft_precise import FtPrecise
 from freqtrade.util.periodic_cache import PeriodicCache
 
@@ -9,6 +10,7 @@ __all__ = [
     'dt_now',
     'dt_ts',
     'dt_utc',
+    'shorten_date',
     'FtPrecise',
     'PeriodicCache',
 ]
